Datastream sats

+++ components/replayengine/db/migrations/001_init.sql
@@ -1,13 +1,13 @@
 CREATE TABLE bars_1sec (
     symbol TEXT,
     t      BIGINT,
-    o      REAL,
-    h      REAL,
-    l      REAL,
-    c      REAL,
+    o      DOUBLE PRECISION,
+    h      DOUBLE PRECISION,
+    l      DOUBLE PRECISION,
+    c      DOUBLE PRECISION,
     n      BIGINT,
-    v      REAL,
-    vw     REAL,
+    v      DOUBLE PRECISION,
+    vw     DOUBLE PRECISION,
     PRIMARY KEY (symbol, t)
 );
